+++ api_server/modules/users/login_helper.js
@@ -195,7 +195,7 @@ class LoginHelper {
 			return;
 		}
 
-		const { isOnPrem, runTimeEnvironment } = this.apiConfig.sharedGeneral;
+		const { isOnPrem, runTimeEnvironment, isProductionCloud } = this.apiConfig.sharedGeneral;
 
 		this.responseData = {
 			user: this.user.getSanitizedObjectForMe({ request: this.request }),	// include me-only attributes
@@ -207,7 +207,9 @@ class LoginHelper {
 					interactiveComponentsEnabled: this.api.config.integrations.slack.interactiveComponentsEnabled
 				}
 			},
-			runtimeEnvironment: isOnPrem ? 'onprem' : runTimeEnvironment
+			isOnPrem,
+			isProductionCloud,
+			runtimeEnvironment: runTimeEnvironment
 		};
 		if (this.apiConfig.broadcastEngine.pubnub && this.apiConfig.broadcastEngine.pubnub.subscribeKey) {
 			this.responseData.pubnubKey = this.apiConfig.broadcastEngine.pubnub.subscribeKey;	// give them the subscribe key for pubnub