+++ api_server/modules/users/register_request.js
@@ -466,7 +466,7 @@ class RegisterRequest extends RestfulRequest {
 			},
 			errors: [
 				'parameterRequired',
-				'usernameNotUnique',
+				//'usernameNotUnique',
 				'exists',
 				'validation',
 				'inviteMismatch',