+++ api_server/modules/users/test/post_user/dont_send_email_test.js
@@ -12,6 +12,7 @@ class DontSendEmailTest extends UnregisteredInviteTest {
 	// form the data for the user update
 	makeUserData (callback) {
 		// add dontSendEmail flag
+		this.noLastInviteType = this.noFirstInviteType = true;
 		super.makeUserData(() => {
 			this.data.dontSendEmail = true;
 			callback();