+++ api_server/modules/users/test/post_user/post_user_request_tester.js
@@ -33,8 +33,8 @@ const TrackingAlreadyInvitedTest = require('./tracking_already_invited_test');
 //const ProviderTrackingTest = require('./provider_tracking_test');
 const NoTrackingTest = require('./no_tracking_test');
 */
-const UsernameResolutionTest = require('./username_resolution_test');
-const ExtendedUsernameResolutionTest = require('./extended_username_resolution_test');
+//const UsernameResolutionTest = require('./username_resolution_test');
+//const ExtendedUsernameResolutionTest = require('./extended_username_resolution_test');
 const UsernameResolutionWithIllegalCharactersTest = require('./username_resolution_with_illegal_characters_test');
 const OnlyAdminsTest = require('./only_admins_test');
 const AdminsCanInviteTest = require('./admins_can_invite_test');
@@ -86,8 +86,8 @@ class PostUserRequestTester {
 		new NoTrackingTest().test();
 		// ProviderTrackingTest().test();
 		*/
-		new UsernameResolutionTest().test();
-		new ExtendedUsernameResolutionTest().test();
+		//new UsernameResolutionTest().test();
+		//new ExtendedUsernameResolutionTest().test();
 		new UsernameResolutionWithIllegalCharactersTest().test();
 		new OnlyAdminsTest().test();
 		new AdminsCanInviteTest().test();