+++ api_server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [8.2.36] - 2021-4-6
+
+### Changed 
+
+- Eliminates check for username uniqueness at registration since it's really a display name
+- Improves logic for determining on-prem vs cloud
+
 ## [8.2.35] - 2021-3-26
 
 ### Fixed